Code/requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
tensorflow==2.11.0
imbalanced-learn==0.10.1
streamlit==1.22.0
matplotlib==3.7.1
seaborn==0.12.2
pillow==9.5.0
h5py==3.8.0
protobuf>=3.9.2,<3.20
